implement deployment settings for vercel and update readme
