+++ components/AppFooter.vue
@@ -1,0 +1,23 @@
+<template>
+  <footer
+    class="mt-16 border-t border-gray-200 dark:border-gray-700 bg-white/50 dark:bg-gray-800/50 backdrop-blur-sm"
+  >
+    <div class="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8 py-6">
+      <div class="text-center text-sm text-gray-600 dark:text-gray-400">
+        <p>
+          Powered by
+          <a
+            href="https://restcountries.com"
+            target="_blank"
+            class="text-blue-600 dark:text-blue-400 hover:underline"
+            >REST Countries API</a
+          >
+        </p>
+      </div>
+    </div>
+  </footer>
+</template>
+
+<script setup lang="ts">
+// No props or logic needed for this simple footer component
+</script>
