+++ package.json
@@ -3,7 +3,7 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "build": "node scripts/build.js",
+    "build": "nuxt build",
     "dev": "node scripts/dev.js",
     "dev:clean": "nuxt dev",
     "generate": "nuxt generate",
