+++ scripts/demo.js
@@ -2,7 +2,7 @@
 
 console.clear()
 
-console.log('\n🎬 \x1b[36m\x1b[1mCountry Explorer\x1b[0m - Interview Demo Guide')
+console.log('\n🎬 \x1b[36m\x1b[1mCountry Explorer\x1b[0m - Application Overview')
 console.log('\x1b[2m═══════════════════════════════════════════════════════\x1b[0m')
 
 console.log('\n🚀 \x1b[1mQuick Start:\x1b[0m')
@@ -10,7 +10,7 @@ console.log('   npm run dev     # Start development server')
 console.log('   npm run build   # Production build')
 console.log('   npm run test    # Run comprehensive test suite')
 
-console.log('\n✨ \x1b[1mKey Features to Showcase:\x1b[0m')
+console.log('\n✨ \x1b[1mKey Features:\x1b[0m')
 console.log('   🔍 Real-time search with intelligent caching')
 console.log('   🌙 Dark/Light mode with smooth transitions')
 console.log('   📱 Fully responsive design (mobile-first)')
@@ -24,7 +24,7 @@ console.log('   • Pinia for state management')
 console.log('   • Vitest for testing (25+ tests)')
 console.log('   • REST Countries API integration')
 
-console.log('\n🎯 \x1b[1mInterview Highlights:\x1b[0m')
+console.log('\n🎯 \x1b[1mTechnical Highlights:\x1b[0m')
 console.log('   ✅ Modern Vue.js development patterns')
 console.log('   ✅ TypeScript for type safety')
 console.log('   ✅ Comprehensive testing strategy')
@@ -32,18 +32,20 @@ console.log('   ✅ Performance optimization techniques')
 console.log('   ✅ Responsive design implementation')
 console.log('   ✅ State management best practices')
 
-console.log('\n📖 \x1b[1mDemo Flow:\x1b[0m')
+console.log('\n📖 \x1b[1mApplication Flow:\x1b[0m')
 console.log('   1. Search for "United" → Real-time filtering')
 console.log('   2. Select USA → Detailed country information')
 console.log('   3. Toggle dark mode → Smooth theme transition')
 console.log('   4. Resize browser → Responsive layout adaptation')
 console.log('   5. Refresh page → Intelligent caching demonstration')
 
-console.log('\n🧪 \x1b[1mTesting:\x1b[0m')
+console.log('\n🧪 \x1b[1mTesting Coverage:\x1b[0m')
 console.log('   • 17 unit tests for store logic')
 console.log('   • 8 integration tests for user flows')
 console.log('   • Error handling and edge cases covered')
 console.log('   • Performance testing with large datasets')
 
 console.log('\n\x1b[2m═══════════════════════════════════════════════════════\x1b[0m')
-console.log('🌟 \x1b[32m\x1b[1mReady to impress! This showcases modern Vue.js expertise.\x1b[0m\n')
+console.log(
+  '🌟 \x1b[32m\x1b[1mA comprehensive Vue.js application showcasing modern development practices.\x1b[0m\n',
+)
